refactor: move HTML UI code from utils.rs to views.rs
